+++ CHANGELOG.md
@@ -18,8 +18,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   `Cargo.lock` is exactly what is expected (without implicit changes at build
   time) but this may end up rejecting builds which were previously passing. To
   get the old behavior back, set `cargoExtraArgs = "";`
-* **Breaking**: `cargoDoc` will no longer install cargo artifacts by default.
-  Set `doInstallCargoArtifacts = true;` to get the old behavior back.
+* **Breaking**: `cargoDoc` and `cargoClippy` will no longer install cargo artifacts
+  by default. Set `doInstallCargoArtifacts = true;` to get the old behavior back.
+* **Breaking**: rustc is now executed with a wrapper script. Therefore providing
+  your own value of `RUSTC_WRAPPER` will no longer work. We do not believe that
+  this is commonly used.
+* Large performance improvement when symlinking artifacts.
 * `cargoDoc` will now install generated documentation in `$out/share/doc`
 * Fixed a bug when testing proc macro crates with `cargoNextest` on macOS.
   ([#376](https://github.com/ipetkov/crane/pull/376))